Add example of dynamic layout generation
This commit adds an initial very simple implementation of re-generating
the layout for the page when the content changes. Currently this is done
monolithically for the whole app at once, but next step is to add
metadata to allow the app to re-generate only small parts of the tree
when they are marked dirty.