Switch back to menu layout
This commit switches back to the previous "menu + text" layout that I
had working with mouse clicks a few commits ago, but using the new
"single component update" framework. This attaches a generation function
to particular parts of the tree, and gives a persistent ID to those
components, and only re-generates that part of the tree when they
change.