Added slider control
This adds the drawing functionality for a slider control. It does not
propperly react to mouse movement right now; it only responds to clicks.
TODO: Make it respond to dragging. Creating this component also involved
generalizing the "roundRect" shader program to allow providing a top and
bottom "tint" or thin highlight and shadow. Maybe should rename from
tint to accent or something like that.