Fix null-terminator bug on Chrome
Previously I was accidentally passing a bunch of null-terminated strings
with the null-terminator byte included to the WebGL API. It seems like
WebGL is ok with this on Safari; it must do its own string length check.
But Chrome complained about uniform block names not being ASCII. This
commit fixes that.